refactor(api, dto, front): remove deprecated use of search/association/:id
