+++ packages/dto/associations/GetAssociationResponseDto.ts
@@ -1,5 +1,5 @@
 import { SuccessResponse, ErrorResponse } from "../shared/ResponseStatus";
-import { Association } from "../search/Association";
+import { Association } from "./Association";
 
 export interface GetAssociationSuccessResponse extends SuccessResponse {
     association: Association,
