+++ packages/dto/index.ts
@@ -3,8 +3,7 @@ export * from "./user/index"
 export * from "./auth/LoginDtoResponse"
 export * from "./auth/ResetPasswordDtoResponse"
 export * from "./auth/SignupDtoResponse"
-export * from "./search/Association"
-export * from "./search/AssociationDtoResponse"
+export * from "./associations/Association"
 export * from "./search/EtablissementDtoResponse"
 export * from "./search/DemandeSubvention"
 export * from "./search/Etablissement"
