--- packages/dto/search/AssociationDtoResponse.ts
@@ -1,7 +1,0 @@
-import { Association } from "../index";
-
-export default interface AssociationDtoResponse {
-    success: boolean;
-    association?: Association,
-    message?: string
-}
